instead of using create i switched to has table logic

+++ Backend/database/migrations/2025_01_22_173242_create_scheduling_table.php
@@ -13,21 +13,23 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('scheduling', function (Blueprint $table) {
-            $table->string('schedcodeID', 200)->primary();
-            $table->integer('loginID')->nullable()->index('PCY');
-            $table->integer('subblockedID')->nullable();
-            $table->time('start')->nullable();
-            $table->time('end')->nullable();
-            $table->string('day', 200)->nullable();
-            $table->string('room', 200)->nullable();
-            $table->integer('maxstud')->nullable();
-            $table->integer('stat')->default(0);
-            $table->timestamp('dtencode')->nullable();
-            $table->timestamp('dtSubmit')->nullable();
-            $table->timestamp('dtLock')->nullable();
-            $table->integer('stat_teach')->default(0);
-        });
+        if (!Schema::hasTable('scheduling')) {
+            Schema::create('scheduling', function (Blueprint $table) {
+                $table->string('schedcodeID', 200)->primary();
+                $table->integer('loginID')->nullable()->index('PCY');
+                $table->integer('subblockedID')->nullable();
+                $table->time('start')->nullable();
+                $table->time('end')->nullable();
+                $table->string('day', 200)->nullable();
+                $table->string('room', 200)->nullable();
+                $table->integer('maxstud')->nullable();
+                $table->integer('stat')->default(0);
+                $table->timestamp('dtencode')->nullable();
+                $table->timestamp('dtSubmit')->nullable();
+                $table->timestamp('dtLock')->nullable();
+                $table->integer('stat_teach')->default(0);
+            });
+        }
     }
 
     /**
@@ -37,6 +39,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('scheduling');
+        if (Schema::hasTable('scheduling')) {
+            Schema::dropIfExists('scheduling');
+        }
     }
 };
